+++ NEWS.md
@@ -1,5 +1,6 @@
 # sclet 0.0.3.001
 
++ Import 'Matrix' for calculate stats of SparseMatrix (2024-10-07, Mon)
 + re-export `aplot::plot_list()` (2024-10-06, Sun)
 
 # sclet 0.0.3